split manager into multiple files
